pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "uc-catalog-mcp"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "psycopg[binary]>=3.1",
    "psycopg2-binary>=2.9",
    "httpx>=0.27",
    "pgvector>=0.3",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "databricks-sdk>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "sync*"]

[tool.ruff]
line-length = 100
